requirements.txt
asyncclick
anyio
progress
pillow
bleak
fastapi
uvicorn
httpx
python-multipart